make Rectangle extends LatLonGeometry
