+++ lucene/core/src/java/org/apache/lucene/geo/Rectangle.java
@@ -33,7 +33,7 @@ import static org.apache.lucene.util.SloppyMath.asin;
 import static org.apache.lucene.util.SloppyMath.cos;
 
 /** Represents a lat/lon rectangle. */
-public class Rectangle {
+public class Rectangle extends LatLonGeometry {
   /** maximum longitude value (in degrees) */
   public final double minLat;
   /** minimum longitude value (in degrees) */
@@ -58,6 +58,11 @@ public class Rectangle {
     assert maxLat >= minLat;
 
     // NOTE: cannot assert maxLon >= minLon since this rect could cross the dateline
+  }
+
+  @Override
+  protected Component2D toComponent2D() {
+    return Rectangle2D.create(this);
   }
 
   @Override
