+++ lucene/core/src/java/org/apache/lucene/geo/Rectangle2D.java
@@ -14,364 +14,158 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package org.apache.lucene.geo;
 
-import java.util.Arrays;
 import java.util.Objects;
 
-import org.apache.lucene.index.PointValues.Relation;
-import org.apache.lucene.util.NumericUtils;
+import org.apache.lucene.index.PointValues;
 
-import static java.lang.Integer.BYTES;
-import static org.apache.lucene.geo.GeoEncodingUtils.MAX_LON_ENCODED;
-import static org.apache.lucene.geo.GeoEncodingUtils.MIN_LON_ENCODED;
 import static org.apache.lucene.geo.GeoEncodingUtils.encodeLatitude;
 import static org.apache.lucene.geo.GeoEncodingUtils.encodeLatitudeCeil;
 import static org.apache.lucene.geo.GeoEncodingUtils.encodeLongitude;
 import static org.apache.lucene.geo.GeoEncodingUtils.encodeLongitudeCeil;
+import static org.apache.lucene.geo.GeoEncodingUtils.decodeLatitude;
+import static org.apache.lucene.geo.GeoEncodingUtils.decodeLongitude;
+import static org.apache.lucene.geo.GeoUtils.MAX_LON_INCL;
+import static org.apache.lucene.geo.GeoUtils.MIN_LON_INCL;
 import static org.apache.lucene.geo.GeoUtils.orient;
 
 /**
- * 2D rectangle implementation containing geo spatial logic.
- *
- * @lucene.internal
+ * 2D rectangle implementation containing cartesian spatial logic.
  */
-public class Rectangle2D {
-  protected final byte[] bbox;
-  private final byte[] west;
-  protected final int minX;
-  protected final int maxX;
-  protected final int minY;
-  protected final int maxY;
+final class Rectangle2D implements Component2D {
 
-  protected Rectangle2D(double minLat, double maxLat, double minLon, double maxLon) {
-    this.bbox = new byte[4 * BYTES];
-    int minXenc = encodeLongitudeCeil(minLon);
-    int maxXenc = encodeLongitude(maxLon);
-    int minYenc = encodeLatitudeCeil(minLat);
-    int maxYenc = encodeLatitude(maxLat);
-    if (minYenc > maxYenc) {
-      minYenc = maxYenc;
-    }
-    this.minY = minYenc;
-    this.maxY = maxYenc;
+  private final double minX;
+  private final double maxX;
+  private final double minY;
+  private final double maxY;
 
-    if (minLon > maxLon == true) {
-      // crossing dateline is split into east/west boxes
-      this.west = new byte[4 * BYTES];
-      this.minX = minXenc;
-      this.maxX = maxXenc;
-      encode(MIN_LON_ENCODED, this.maxX, this.minY, this.maxY, this.west);
-      encode(this.minX, MAX_LON_ENCODED, this.minY, this.maxY, this.bbox);
-    } else {
-      // encodeLongitudeCeil may cause minX to be > maxX iff
-      // the delta between the longitude < the encoding resolution
-      if (minXenc > maxXenc) {
-        minXenc = maxXenc;
-      }
-      this.west = null;
-      this.minX = minXenc;
-      this.maxX = maxXenc;
-      encode(this.minX, this.maxX, this.minY, this.maxY, bbox);
-    }
+  private Rectangle2D(double minX, double maxX, double minY, double maxY) {
+    this.minX =  minX;
+    this.maxX =  maxX;
+    this.minY =  minY;
+    this.maxY =  maxY;
   }
 
-  protected Rectangle2D(int minX, int maxX, int minY, int maxY) {
-    this.bbox = new byte[4 * BYTES];
-    this.west = null;
-    this.minX = minX;
-    this.maxX = maxX;
-    this.minY = minY;
-    this.maxY = maxY;
-    encode(this.minX, this.maxX, this.minY, this.maxY, bbox);
+  @Override
+  public double getMinX() {
+    return minX;
   }
 
-  /** Builds a Rectangle2D from rectangle */
-  public static Rectangle2D create(Rectangle rectangle) {
-    return new Rectangle2D(rectangle.minLat, rectangle.maxLat, rectangle.minLon, rectangle.maxLon);
+  @Override
+  public double getMaxX() {
+    return maxX;
   }
 
-  public boolean crossesDateline() {
-    return minX > maxX;
+  @Override
+  public double getMinY() {
+    return minY;
   }
 
-  /** Checks if the rectangle contains the provided point **/
-  public boolean queryContainsPoint(int x, int y) {
-    if (this.crossesDateline() == true) {
-      return bboxContainsPoint(x, y, MIN_LON_ENCODED, this.maxX, this.minY, this.maxY)
-          || bboxContainsPoint(x, y, this.minX, MAX_LON_ENCODED, this.minY, this.maxY);
-    }
-    return bboxContainsPoint(x, y, this.minX, this.maxX, this.minY, this.maxY);
+  @Override
+  public double getMaxY() {
+    return maxY;
   }
 
-  /** compare this to a provided range bounding box **/
-  public Relation relateRangeBBox(int minXOffset, int minYOffset, byte[] minTriangle,
-                                  int maxXOffset, int maxYOffset, byte[] maxTriangle) {
-    Relation eastRelation = compareBBoxToRangeBBox(this.bbox,
-        minXOffset, minYOffset, minTriangle, maxXOffset, maxYOffset, maxTriangle);
-    if (this.crossesDateline() && eastRelation == Relation.CELL_OUTSIDE_QUERY) {
-      return compareBBoxToRangeBBox(this.west, minXOffset, minYOffset, minTriangle, maxXOffset, maxYOffset, maxTriangle);
-    }
-    return eastRelation;
+  @Override
+  public boolean contains(double x, double y) {
+    return Component2D.containsPoint(x, y, this.minX, this.maxX, this.minY, this.maxY);
   }
 
-  /** intersects this to a provided range bounding box **/
-  public Relation intersectRangeBBox(int minXOffset, int minYOffset, byte[] minTriangle,
-                                  int maxXOffset, int maxYOffset, byte[] maxTriangle) {
-    Relation eastRelation = intersectBBoxWithRangeBBox(this.bbox,
-        minXOffset, minYOffset, minTriangle, maxXOffset, maxYOffset, maxTriangle);
-    if (this.crossesDateline() && eastRelation == Relation.CELL_OUTSIDE_QUERY) {
-      return intersectBBoxWithRangeBBox(this.west, minXOffset, minYOffset, minTriangle, maxXOffset, maxYOffset, maxTriangle);
+  @Override
+  public PointValues.Relation relate(double minX, double maxX, double minY, double maxY) {
+    if (Component2D.disjoint(this.minX, this.maxX, this.minY, this.maxY, minX, maxX, minY, maxY)) {
+      return PointValues.Relation.CELL_OUTSIDE_QUERY;
     }
-    return eastRelation;
+    if (Component2D.within(minX, maxX, minY, maxY, this.minX, this.maxX, this.minY, this.maxY)) {
+      return PointValues.Relation.CELL_INSIDE_QUERY;
+    }
+    return PointValues.Relation.CELL_CROSSES_QUERY;
   }
 
-  /** Checks if the rectangle intersects the provided triangle **/
-  public boolean intersectsTriangle(int aX, int aY, int bX, int bY, int cX, int cY) {
-    // 1. query contains any triangle points
-    if (queryContainsPoint(aX, aY) || queryContainsPoint(bX, bY) || queryContainsPoint(cX, cY)) {
-      return true;
-    }
+  @Override
+  public PointValues.Relation relateTriangle(double minX, double maxX, double minY, double maxY,
+                                             double ax, double ay, double bx, double by, double cx, double cy) {
 
-    // compute bounding box of triangle
-    int tMinX = StrictMath.min(StrictMath.min(aX, bX), cX);
-    int tMaxX = StrictMath.max(StrictMath.max(aX, bX), cX);
-    int tMinY = StrictMath.min(StrictMath.min(aY, bY), cY);
-    int tMaxY = StrictMath.max(StrictMath.max(aY, bY), cY);
 
-    // 2. check bounding boxes are disjoint
-    if (this.crossesDateline() == true) {
-      if (boxesAreDisjoint(tMinX, tMaxX, tMinY, tMaxY, MIN_LON_ENCODED, this.maxX, this.minY, this.maxY)
-          && boxesAreDisjoint(tMinX, tMaxX, tMinY, tMaxY, this.minX, MAX_LON_ENCODED, this.minY, this.maxY)) {
-        return false;
-      }
-    } else if (tMaxX < minX || tMinX > maxX || tMinY > maxY || tMaxY < minY) {
-      return false;
+    if (Component2D.disjoint(this.minX, this.maxX, this.minY, this.maxY, minX, maxX, minY, maxY)) {
+      return PointValues.Relation.CELL_OUTSIDE_QUERY;
     }
-
-    // 3. check triangle contains any query points
-    if (Tessellator.pointInTriangle(minX, minY, aX, aY, bX, bY, cX, cY)) {
-      return true;
-    } else if (Tessellator.pointInTriangle(maxX, minY, aX, aY, bX, bY, cX, cY)) {
-      return true;
-    } else if (Tessellator.pointInTriangle(maxX, maxY, aX, aY, bX, bY, cX, cY)) {
-      return true;
-    } else if (Tessellator.pointInTriangle(minX, maxY, aX, aY, bX, bY, cX, cY)) {
-      return true;
+    int edgesContain = numberOfCorners(ax, ay, bx, by, cx, cy);
+    if (edgesContain == 3) {
+      return PointValues.Relation.CELL_INSIDE_QUERY;
+    } else if (edgesContain != 0) {
+      return PointValues.Relation.CELL_CROSSES_QUERY;
+    } else if (Component2D.pointInTriangle(minX, maxX, minY, maxY, this.minX, this.minY,ax, ay, bx, by, cx, cy)
+        || edgesIntersect(ax, ay, bx, by)
+        || edgesIntersect(bx, by, cx, cy)
+        || edgesIntersect(cx, cy, ax, ay)) {
+      return PointValues.Relation.CELL_CROSSES_QUERY;
     }
-
-    // 4. last ditch effort: check crossings
-    if (queryIntersects(aX, aY, bX, bY, cX, cY)) {
-      return true;
-    }
-    return false;
+    return PointValues.Relation.CELL_OUTSIDE_QUERY;
   }
 
-  /** Returns the Within relation to the provided triangle */
-  public Component2D.WithinRelation withinTriangle(int ax, int ay, boolean ab, int bx, int by, boolean bc, int cx, int cy, boolean ca) {
-    if (this.crossesDateline() == true) {
-      throw new IllegalArgumentException("withinTriangle is not supported for rectangles crossing the date line");
-    }
+  @Override
+  public WithinRelation withinTriangle(double minX, double maxX, double minY, double maxY,
+                                       double ax, double ay, boolean ab, double bx, double by, boolean bc, double cx, double cy, boolean ca) {
     // Short cut, lines and points cannot contain a bbox
     if ((ax == bx && ay == by) || (ax == cx && ay == cy) || (bx == cx && by == cy)) {
-      return Component2D.WithinRelation.DISJOINT;
+      return WithinRelation.DISJOINT;
     }
-    // Compute bounding box of triangle
-    int tMinX = StrictMath.min(StrictMath.min(ax, bx), cx);
-    int tMaxX = StrictMath.max(StrictMath.max(ax, bx), cx);
-    int tMinY = StrictMath.min(StrictMath.min(ay, by), cy);
-    int tMaxY = StrictMath.max(StrictMath.max(ay, by), cy);
+
     // Bounding boxes disjoint?
-    if (boxesAreDisjoint(tMinX, tMaxX, tMinY, tMaxY, minX, maxX, minY, maxY)) {
-      return Component2D.WithinRelation.DISJOINT;
+    if (Component2D.disjoint(this.minX, this.maxX, this.minY, this.maxY, minX, maxX, minY, maxY)) {
+      return WithinRelation.DISJOINT;
     }
+
     // Points belong to the shape so if points are inside the rectangle then it cannot be within.
-    if (bboxContainsPoint(ax, ay, minX, maxX, minY, maxY) ||
-        bboxContainsPoint(bx, by, minX, maxX, minY, maxY) ||
-        bboxContainsPoint(cx, cy, minX, maxX, minY, maxY)) {
-      return Component2D.WithinRelation.NOTWITHIN;
+    if (contains(ax, ay) || contains(bx, by) || contains(cx, cy)) {
+      return WithinRelation.NOTWITHIN;
     }
     // If any of the edges intersects an edge belonging to the shape then it cannot be within.
-    Component2D.WithinRelation relation = Component2D.WithinRelation.DISJOINT;
-    if (edgeIntersectsBox(ax, ay, bx, by, minX, maxX, minY, maxY) == true) {
+    WithinRelation relation = WithinRelation.DISJOINT;
+    if (edgesIntersect(ax, ay, bx, by) == true) {
       if (ab == true) {
-        return Component2D.WithinRelation.NOTWITHIN;
+        return WithinRelation.NOTWITHIN;
       } else {
-        relation = Component2D.WithinRelation.CANDIDATE;
+        relation = WithinRelation.CANDIDATE;
       }
     }
-    if (edgeIntersectsBox(bx, by, cx, cy, minX, maxX, minY, maxY) == true) {
+    if (edgesIntersect(bx, by, cx, cy) == true) {
       if (bc == true) {
-        return Component2D.WithinRelation.NOTWITHIN;
+        return WithinRelation.NOTWITHIN;
       } else {
-        relation = Component2D.WithinRelation.CANDIDATE;
+        relation = WithinRelation.CANDIDATE;
       }
     }
 
-    if (edgeIntersectsBox(cx, cy, ax, ay, minX, maxX, minY, maxY) == true) {
+    if (edgesIntersect(cx, cy, ax, ay) == true) {
       if (ca == true) {
-        return Component2D.WithinRelation.NOTWITHIN;
+        return WithinRelation.NOTWITHIN;
       } else {
-        relation = Component2D.WithinRelation.CANDIDATE;
+        relation = WithinRelation.CANDIDATE;
       }
     }
     // If any of the rectangle edges crosses a triangle edge that does not belong to the shape
     // then it is a candidate for within
-    if (relation == Component2D.WithinRelation.CANDIDATE) {
-      return Component2D.WithinRelation.CANDIDATE;
+    if (relation == WithinRelation.CANDIDATE) {
+      return WithinRelation.CANDIDATE;
     }
     // Check if shape is within the triangle
-    if (Tessellator.pointInTriangle(minX, minY, ax, ay, bx, by, cx, cy)) {
-      return Component2D.WithinRelation.CANDIDATE;
+    if (Component2D.pointInTriangle(minX, maxX, minY, maxY, this.minX, this.minY, ax, ay, bx, by, cx, cy)) {
+      return WithinRelation.CANDIDATE;
     }
     return relation;
   }
 
-  /** Checks if the rectangle contains the provided triangle **/
-  public boolean containsTriangle(int ax, int ay, int bx, int by, int cx, int cy) {
-    if (this.crossesDateline() == true) {
-      return bboxContainsTriangle(ax, ay, bx, by, cx, cy, MIN_LON_ENCODED, this.maxX, this.minY, this.maxY)
-          || bboxContainsTriangle(ax, ay, bx, by, cx, cy, this.minX, MAX_LON_ENCODED, this.minY, this.maxY);
-    }
-    return bboxContainsTriangle(ax, ay, bx, by, cx, cy, minX, maxX, minY, maxY);
-  }
-
-  /**
-   * static utility method to compare a bbox with a range of triangles (just the bbox of the triangle collection)
-   **/
-  private static Relation compareBBoxToRangeBBox(final byte[] bbox,
-                                                 int minXOffset, int minYOffset, byte[] minTriangle,
-                                                 int maxXOffset, int maxYOffset, byte[] maxTriangle) {
-    // check bounding box (DISJOINT)
-    if (disjoint(bbox, minXOffset, minYOffset, minTriangle, maxXOffset, maxYOffset, maxTriangle)) {
-      return Relation.CELL_OUTSIDE_QUERY;
-    }
-
-    if (Arrays.compareUnsigned(minTriangle, minXOffset, minXOffset + BYTES, bbox, BYTES, 2 * BYTES) >= 0 &&
-        Arrays.compareUnsigned(maxTriangle, maxXOffset, maxXOffset + BYTES, bbox, 3 * BYTES, 4 * BYTES) <= 0 &&
-        Arrays.compareUnsigned(minTriangle, minYOffset, minYOffset + BYTES, bbox, 0, BYTES) >= 0 &&
-        Arrays.compareUnsigned(maxTriangle, maxYOffset, maxYOffset + BYTES, bbox, 2 * BYTES, 3 * BYTES) <= 0) {
-      return Relation.CELL_INSIDE_QUERY;
-    }
-
-    return Relation.CELL_CROSSES_QUERY;
-  }
-
-  /**
-   * static utility method to compare a bbox with a range of triangles (just the bbox of the triangle collection)
-   * for intersection
-   **/
-  private static Relation intersectBBoxWithRangeBBox(final byte[] bbox,
-                                                     int minXOffset, int minYOffset, byte[] minTriangle,
-                                                     int maxXOffset, int maxYOffset, byte[] maxTriangle) {
-    // check bounding box (DISJOINT)
-    if (disjoint(bbox, minXOffset, minYOffset, minTriangle, maxXOffset, maxYOffset, maxTriangle)) {
-      return Relation.CELL_OUTSIDE_QUERY;
-    }
-
-    if (Arrays.compareUnsigned(minTriangle, minXOffset, minXOffset + BYTES, bbox, BYTES, 2 * BYTES) >= 0 &&
-        Arrays.compareUnsigned(minTriangle, minYOffset, minYOffset + BYTES, bbox, 0, BYTES) >= 0 ) {
-      if (Arrays.compareUnsigned(maxTriangle, minXOffset, minXOffset + BYTES, bbox, 3 * BYTES, 4 * BYTES) <= 0 &&
-          Arrays.compareUnsigned(maxTriangle, maxYOffset, maxYOffset + BYTES, bbox, 2 * BYTES, 3 * BYTES) <= 0) {
-        return Relation.CELL_INSIDE_QUERY;
-      }
-      if (Arrays.compareUnsigned(maxTriangle, maxXOffset, maxXOffset + BYTES, bbox, 3 * BYTES, 4 * BYTES) <= 0 &&
-          Arrays.compareUnsigned(maxTriangle, minYOffset, minYOffset + BYTES, bbox, 2 * BYTES, 3 * BYTES) <= 0) {
-        return Relation.CELL_INSIDE_QUERY;
-      }
-    }
-
-    if (Arrays.compareUnsigned(maxTriangle, maxXOffset, maxXOffset + BYTES, bbox, 3 * BYTES, 4 * BYTES) <= 0 &&
-        Arrays.compareUnsigned(maxTriangle, maxYOffset, maxYOffset + BYTES, bbox, 2 * BYTES, 3 * BYTES) <= 0 ) {
-      if (Arrays.compareUnsigned(minTriangle, minXOffset, minXOffset + BYTES, bbox, BYTES, 2 * BYTES) >= 0 &&
-          Arrays.compareUnsigned(minTriangle, maxYOffset, maxYOffset + BYTES, bbox, 0, BYTES) >= 0) {
-        return Relation.CELL_INSIDE_QUERY;
-      }
-      if (Arrays.compareUnsigned(minTriangle, maxXOffset, maxXOffset + BYTES, bbox, BYTES, 2 * BYTES) >= 0 &&
-          Arrays.compareUnsigned(minTriangle, minYOffset, minYOffset + BYTES, bbox, 0, BYTES) >= 0) {
-        return Relation.CELL_INSIDE_QUERY;
-      }
-    }
-
-    return Relation.CELL_CROSSES_QUERY;
-  }
-
-  /**
-   * static utility method to check a bbox is disjoint with a range of triangles
-   **/
-  private static boolean disjoint(final byte[] bbox,
-                               int minXOffset, int minYOffset, byte[] minTriangle,
-                               int maxXOffset, int maxYOffset, byte[] maxTriangle) {
-      return Arrays.compareUnsigned(minTriangle, minXOffset, minXOffset + BYTES, bbox, 3 * BYTES, 4 * BYTES) > 0 ||
-          Arrays.compareUnsigned(maxTriangle, maxXOffset, maxXOffset + BYTES, bbox, BYTES, 2 * BYTES) < 0 ||
-          Arrays.compareUnsigned(minTriangle, minYOffset, minYOffset + BYTES, bbox, 2 * BYTES, 3 * BYTES) > 0 ||
-          Arrays.compareUnsigned(maxTriangle, maxYOffset, maxYOffset + BYTES, bbox, 0, BYTES) < 0;
-  }
-
-  /**
-   * encodes a bounding box into the provided byte array
-   */
-  private static void encode(final int minX, final int maxX, final int minY, final int maxY, byte[] b) {
-    if (b == null) {
-      b = new byte[4 * BYTES];
-    }
-    NumericUtils.intToSortableBytes(minY, b, 0);
-    NumericUtils.intToSortableBytes(minX, b, BYTES);
-    NumericUtils.intToSortableBytes(maxY, b, 2 * BYTES);
-    NumericUtils.intToSortableBytes(maxX, b, 3 * BYTES);
-  }
-
-  /** returns true if the query intersects the provided triangle (in encoded space) */
-  private boolean queryIntersects(int ax, int ay, int bx, int by, int cx, int cy) {
-    // check each edge of the triangle against the query
-    if (edgeIntersectsQuery(ax, ay, bx, by) ||
-        edgeIntersectsQuery(bx, by, cx, cy) ||
-        edgeIntersectsQuery(cx, cy, ax, ay)) {
-      return true;
-    }
-    return false;
-  }
-
-  /** returns true if the edge (defined by (ax, ay) (bx, by)) intersects the query */
-  private boolean edgeIntersectsQuery(int ax, int ay, int bx, int by) {
-    if (this.crossesDateline() == true) {
-      return edgeIntersectsBox(ax, ay, bx, by, MIN_LON_ENCODED, this.maxX, this.minY, this.maxY)
-          || edgeIntersectsBox(ax, ay, bx, by, this.minX, MAX_LON_ENCODED, this.minY, this.maxY);
-    }
-    return edgeIntersectsBox(ax, ay, bx, by, this.minX, this.maxX, this.minY, this.maxY);
-  }
-
-  /** static utility method to check if a bounding box contains a point */
-  private static boolean bboxContainsPoint(int x, int y, int minX, int maxX, int minY, int maxY) {
-    return (x < minX || x > maxX || y < minY || y > maxY) == false;
-  }
-
-  /** static utility method to check if a bounding box contains a triangle */
-  private static boolean bboxContainsTriangle(int ax, int ay, int bx, int by, int cx, int cy,
-                                             int minX, int maxX, int minY, int maxY) {
-    return bboxContainsPoint(ax, ay, minX, maxX, minY, maxY)
-        && bboxContainsPoint(bx, by, minX, maxX, minY, maxY)
-        && bboxContainsPoint(cx, cy, minX, maxX, minY, maxY);
-  }
-
-  /** returns true if the edge (defined by (ax, ay) (bx, by)) intersects the query */
-  private static boolean edgeIntersectsBox(int ax, int ay, int bx, int by,
-                                           int minX, int maxX, int minY, int maxY) {
+  private  boolean edgesIntersect(double ax, double ay, double bx, double by) {
     // shortcut: if edge is a point (occurs w/ Line shapes); simply check bbox w/ point
     if (ax == bx && ay == by) {
-      return Rectangle.containsPoint(ay, ax, minY, maxY, minX, maxX);
-    }
-
-    // shortcut: check if either of the end points fall inside the box
-    if (bboxContainsPoint(ax, ay, minX, maxX, minY, maxY) ||
-        bboxContainsPoint(bx, by, minX, maxX, minY, maxY)) {
-      return true;
+      return false;
     }
 
     // shortcut: check bboxes of edges are disjoint
-    if (boxesAreDisjoint(Math.min(ax, bx), Math.max(ax, bx), Math.min(ay, by), Math.max(ay, by),
-        minX, maxX, minY, maxY)) {
+    if ( Math.max(ax, bx) < minX || Math.min(ax, bx) > maxX || Math.min(ay, by) > maxY || Math.max(ay, by) < minY) {
       return false;
     }
 
@@ -401,10 +195,18 @@ public class Rectangle2D {
     return false;
   }
 
-  /** utility method to check if two boxes are disjoint */
-  private static boolean boxesAreDisjoint(final int aMinX, final int aMaxX, final int aMinY, final int aMaxY,
-                                         final int bMinX, final int bMaxX, final int bMinY, final int bMaxY) {
-    return (aMaxX < bMinX || aMinX > bMaxX || aMaxY < bMinY || aMinY > bMaxY);
+  private int numberOfCorners(double ax, double ay, double bx, double by, double cx, double cy) {
+    int containsCount = 0;
+    if (contains(ax, ay)) {
+      containsCount++;
+    }
+    if (contains(bx, by)) {
+      containsCount++;
+    }
+    if (contains(cx, cy)) {
+      containsCount++;
+    }
+    return containsCount;
   }
 
   @Override
@@ -415,16 +217,63 @@ public class Rectangle2D {
     return minX == that.minX &&
         maxX == that.maxX &&
         minY == that.minY &&
-        maxY == that.maxY &&
-        Arrays.equals(bbox, that.bbox) &&
-        Arrays.equals(west, that.west);
+        maxY == that.maxY;
   }
 
   @Override
   public int hashCode() {
     int result = Objects.hash(minX, maxX, minY, maxY);
-    result = 31 * result + Arrays.hashCode(bbox);
-    result = 31 * result + Arrays.hashCode(west);
     return result;
+  }
+
+  @Override
+  public String toString() {
+    final StringBuilder sb = new StringBuilder();
+    sb.append("XYRectangle(x=");
+    sb.append(minX);
+    sb.append(" TO ");
+    sb.append(maxX);
+    sb.append(" y=");
+    sb.append(minY);
+    sb.append(" TO ");
+    sb.append(maxY);
+    sb.append(")");
+    return sb.toString();
+  }
+
+  /** create a component2D from the provided XY rectangle */
+  static Component2D create(XYRectangle rectangle) {
+    return new Rectangle2D(rectangle.minX, rectangle.maxX, rectangle.minY, rectangle.maxY);
+  }
+
+  private static double MIN_LON_INCL_QUANTIZE = decodeLongitude(encodeLongitude(MIN_LON_INCL));
+  private static double MAX_LON_INCL_QUANTIZE = decodeLongitude(encodeLongitude(MAX_LON_INCL));
+
+  /** create a component2D from the provided LatLon rectangle */
+  static Component2D create(Rectangle rectangle) {
+    // need to quantize!
+    double qMinLat = decodeLatitude(encodeLatitudeCeil(rectangle.minLat));
+    double qMaxLat = decodeLatitude(encodeLatitude(rectangle.maxLat));
+    double qMinLon = decodeLongitude(encodeLongitudeCeil(rectangle.minLon));
+    double qMaxLon = decodeLongitude(encodeLongitude(rectangle.maxLon));
+    if (qMinLat > qMaxLat) {
+      // encodeLatitudeCeil may cause minY to be > maxY iff
+      // the delta between the longitude < the encoding resolution
+      qMinLat = qMaxLat;
+    }
+    if (rectangle.minLon > rectangle.maxLon) {
+      // for rectangles that cross the dateline we need to create two components
+      Component2D[] components = new Component2D[2];
+      components[0] = new Rectangle2D(MIN_LON_INCL_QUANTIZE, qMaxLon, qMinLat, qMaxLat);
+      components[1] = new Rectangle2D(qMinLon, MAX_LON_INCL_QUANTIZE, qMinLat, qMaxLat);
+      return ComponentTree.create(components);
+    } else {
+      // encodeLongitudeCeil may cause minX to be > maxX iff
+      // the delta between the longitude < the encoding resolution
+      if (qMinLon > qMaxLon) {
+        qMinLon = qMaxLon;
+      }
+      return new Rectangle2D(qMinLon, qMaxLon, qMinLat, qMaxLat);
+    }
   }
 }
