+++ lucene/core/src/test/org/apache/lucene/document/BaseLatLonShapeTestCase.java
@@ -26,8 +26,13 @@ import org.apache.lucene.geo.LatLonGeometry;
 import org.apache.lucene.geo.Line;
 import org.apache.lucene.geo.Polygon;
 import org.apache.lucene.geo.Rectangle;
+import org.apache.lucene.index.IndexReader;
+import org.apache.lucene.index.RandomIndexWriter;
+import org.apache.lucene.search.IndexSearcher;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.QueryUtils;
+import org.apache.lucene.store.Directory;
+import org.apache.lucene.util.IOUtils;
 import org.apache.lucene.util.TestUtil;
 
 import static org.apache.lucene.geo.GeoEncodingUtils.decodeLatitude;
@@ -181,6 +186,46 @@ public abstract class BaseLatLonShapeTestCase extends BaseShapeTestCase {
     } else {
       QueryUtils.checkUnequal(q1, q5);
     }
+  }
+
+  public void testBoundingBoxQueriesEquivalence() throws Exception {
+    int numShapes = atLeast(20);
+
+    Directory dir = newDirectory();
+    RandomIndexWriter w = new RandomIndexWriter(random(), dir);
+
+    for (int  i =0; i < numShapes; i++) {
+      indexRandomShapes(w.w, nextShape());
+    }
+    if (random().nextBoolean()) {
+      w.forceMerge(1);
+    }
+
+    ///// search //////
+    IndexReader reader = w.getReader();
+    w.close();
+    IndexSearcher searcher = newSearcher(reader);
+
+    Rectangle box = GeoTestUtil.nextBox();
+
+    Query q1 = LatLonShape.newBoxQuery(FIELD_NAME, QueryRelation.INTERSECTS, box.minLat, box.maxLat, box.minLon, box.maxLon);
+    Query q2 = new LatLonShapeQuery(FIELD_NAME, QueryRelation.INTERSECTS, box);
+    assertEquals(searcher.count(q1), searcher.count(q2));
+    q1 = LatLonShape.newBoxQuery(FIELD_NAME, QueryRelation.WITHIN, box.minLat, box.maxLat, box.minLon, box.maxLon);
+    q2 = new LatLonShapeQuery(FIELD_NAME, QueryRelation.WITHIN, box);
+    assertEquals(searcher.count(q1), searcher.count(q2));
+    q1 = LatLonShape.newBoxQuery(FIELD_NAME, QueryRelation.CONTAINS, box.minLat, box.maxLat, box.minLon, box.maxLon);
+    if (box.crossesDateline()) {
+      q2 = LatLonShape.newGeometryQuery(FIELD_NAME, QueryRelation.CONTAINS, box);
+    } else {
+      q2 = new LatLonShapeQuery(FIELD_NAME, QueryRelation.CONTAINS, box);
+    }
+    assertEquals(searcher.count(q1), searcher.count(q2));
+    q1 = LatLonShape.newBoxQuery(FIELD_NAME, QueryRelation.DISJOINT, box.minLat, box.maxLat, box.minLon, box.maxLon);
+    q2 = new LatLonShapeQuery(FIELD_NAME, QueryRelation.DISJOINT, box);
+    assertEquals(searcher.count(q1), searcher.count(q2));
+
+    IOUtils.close(w, reader, dir);
   }
 
   /** factory method to create a new polygon query */
